LPS-116770 Move undo shortcuts to ShortcutManager
