+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/components/undo/Undo.js
@@ -13,65 +13,13 @@
  */
 
 import ClayButton, {ClayButtonWithIcon} from '@clayui/button';
-import {useEventListener} from 'frontend-js-react-web';
-import {closest} from 'metal-dom';
 import PropTypes from 'prop-types';
 import React from 'react';
 
-import {Z_KEYCODE} from '../../config/constants/keycodes';
 import {useSelector} from '../../store/index';
 import UndoHistory from './UndoHistory';
 
-const isTextElement = (element) => {
-	return (
-		(element.tagName === 'INPUT' && element.type === 'text') ||
-		element.tagName === 'TEXTAREA'
-	);
-};
-
-const isCommentsAlloyEditor = (element) => {
-	return (
-		element.classList.contains('alloy-editor') &&
-		element.parentElement.classList.contains('alloy-editor-container') &&
-		closest(element, '.page-editor__sidebar')
-	);
-};
-
-const isWithinIframe = () => {
-	return window.top !== window.self;
-};
-
-const isUndoCombination = (event) => {
-	const ctrlOrMeta =
-		(event.ctrlKey && !event.metaKey) || (!event.ctrlKey && event.metaKey);
-
-	return event.keyCode === Z_KEYCODE && ctrlOrMeta && !event.altKey;
-};
-
 export default function Undo({onRedo = () => {}, onUndo = () => {}}) {
-	useEventListener(
-		'keydown',
-		(event) => {
-			if (
-				isUndoCombination(event) &&
-				!isTextElement(event.target) &&
-				!isCommentsAlloyEditor(event.target) &&
-				!isWithinIframe()
-			) {
-				event.preventDefault();
-
-				if (event.shiftKey) {
-					onRedo();
-				}
-				else {
-					onUndo();
-				}
-			}
-		},
-		true,
-		window
-	);
-
 	const undoHistory = useSelector((state) => state.undoHistory);
 	const redoHistory = useSelector((state) => state.redoHistory);
